pyproject.toml
[project]
name = "sql-query-debugger"
version = "1.0.0"
description = "An RL environment where an AI agent debugs broken SQL queries"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
